pyproject.toml
[tool.poetry]
name = "mvp-data-pipeline-aws"
version = "0.1.0"
description = "Template for MVP Data Pipeline in AWS"
authors = ["Erik Munkby <erik.munkby@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.core" = "^1.70.0"
"aws-cdk.aws-lambda" = "^1.70.0"
"aws-cdk.aws-apigateway" = "^1.70.0"
"aws-cdk.aws-kinesisfirehose" = "^1.70.0"
"aws-cdk.aws-s3" = "^1.70.0"
"aws-cdk.aws-iam" = "^1.70.0"

[tool.poetry.dev-dependencies]
pandas = "^1.1.3"
jupyter = "^1.0.0"
ipykernel = "^5.3.4"
boto3 = "^1.16.6"
requests = "^2.24.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
